+++ source/src/library/uFuture.cc
@@ -7,8 +7,8 @@
 // Author           : Peter A. Buhr
 // Created On       : Mon Jun  3 18:06:58 2019
 // Last Modified By : Peter A. Buhr
-// Last Modified On : Wed Aug 26 20:42:34 2020
-// Update Count     : 8
+// Last Modified On : Sun Jan 29 20:45:38 2023
+// Update Count     : 9
 //
 // This  library is free  software; you  can redistribute  it and/or  modify it
 // under the terms of the GNU Lesser General Public License as published by the
@@ -27,7 +27,7 @@
 #define __U_KERNEL__
 #include <uC++.h>
 #include <uFuture.h>
-unsigned int uExecutor::next = 0;			// demultiplex across worker buffers
+size_t uExecutor::next = 0;								// demultiplex across worker buffers
 
 // Local Variables: //
 // compile-command: "make install" //
